refactor(frontend): extract commons comp from imdemniteLicenciement
