+++ packages/code-du-travail-frontend/src/outils/IndemniteLicenciement/ccn/0016/Categorie.js
@@ -2,8 +2,8 @@ import React from "react";
 import PropTypes from "prop-types";
 import { Field } from "react-final-form";
 
-import { required } from "../../validators";
-import { Label, SectionTitle } from "../../stepStyles";
+import { required } from "../../../common/validators";
+import { Label, SectionTitle } from "../../../common/stepStyles";
 
 export const OUVRIER = "ouvrier";
 export const TAM = "tam";
