+++ packages/code-du-travail-frontend/src/outils/IndemniteLicenciement/ccn/0016/CategoriePeriod.js
@@ -1,10 +1,10 @@
 import React from "react";
-import { SectionTitle, Input, InlineError } from "../../stepStyles";
+import { SectionTitle, Input, InlineError } from "../../../common/stepStyles";
 import { Field } from "react-final-form";
 import styled from "styled-components";
 import { theme } from "@cdt/ui";
 
-import { isNumber } from "../../validators";
+import { isNumber } from "../../../common/validators";
 
 function CategoriePeriod() {
   return (
