+++ packages/code-du-travail-frontend/src/outils/IndemniteLicenciement/ccn/2120/indemnite.js
@@ -1,4 +1,4 @@
-import { sum, round } from "../../math";
+import { sum, round } from "../../../common/math";
 import { max, min, isAfter, differenceInMonths } from "date-fns";
 import {
   CADRE,
