+++ packages/code-du-travail-frontend/src/outils/IndemniteLicenciement/components/SalaireTempsPartiel.js
@@ -4,8 +4,8 @@ import { Field } from "react-final-form";
 import { FieldArray } from "react-final-form-arrays";
 import { OnChange } from "react-final-form-listeners";
 import { Button, theme } from "@cdt/ui";
-import { Input, InlineError } from "../stepStyles";
-import { isNumber } from "../validators";
+import { Input, InlineError } from "../../common/stepStyles";
+import { isNumber } from "../../common/validators";
 
 function SalaireTempsPartiel({ name, onChange }) {
   return (
