+++ packages/code-du-travail-frontend/src/outils/IndemniteLicenciement/components/SalaireTempsPlein.js
@@ -5,8 +5,8 @@ import { FieldArray } from "react-final-form-arrays";
 import styled from "styled-components";
 import { Table as UITable, theme } from "@cdt/ui";
 
-import { Input, InlineError } from "../stepStyles";
-import { isNumber } from "../validators";
+import { Input, InlineError } from "../../common/stepStyles";
+import { isNumber } from "../../common/validators";
 
 function SalaireTempsPlein({ name }) {
   return (
