+++ packages/code-du-travail-frontend/src/outils/IndemniteLicenciement/steps/Indemnite.js
@@ -5,7 +5,7 @@ import { Button, theme } from "@cdt/ui";
 import styled from "styled-components";
 
 import { branches } from "../branches";
-import { Label, SectionTitle } from "../stepStyles";
+import { Label, SectionTitle } from "../../common/stepStyles";
 import { getIndemniteFromFinalForm } from "../indemnite";
 import { IndemniteLegale } from "../components/IndemniteLegale";
