+++ packages/code-du-travail-frontend/src/outils/common/__tests__/StepItems.test.js
