+++ packages/code-du-travail-frontend/src/outils/common/__tests__/validators.test.js
