+++ packages/code-du-travail-frontend/src/outils/common/math.js
