+++ packages/code-du-travail-frontend/src/outils/common/stepStyles.js
