+++ packages/code-du-travail-frontend/src/outils/common/validators.js
